+++ components/home/why-grouparoo/icons/BriefcaseCheck.tsx
