+++ components/home/why-grouparoo/icons/MoneyBill.tsx
