+++ pages/index.tsx
@@ -20,7 +20,7 @@ import { Testimonials } from "../components/home/testimonials/Testimonials";
 import { HowGrouparooWorks } from "../components/home/how-it-works/HowGrouparooWorks";
 import styles from "./index.module.scss";
 import { Integrations } from "../components/home/integrations/Integrations";
-import { WhyGrouparoo } from "../components/home/WhyGrouparoo/WhyGrouparoo";
+import { WhyGrouparoo } from "../components/home/why-grouparoo/WhyGrouparoo";
 
 const getTagline = (pluginNames: string[]) =>
   `Stop writing code to sync data to ${
